% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_fp.R
\name{add_fp}
\alias{add_fp}
\title{Adds informations about the formation}
\usage{
add_fp(tab1, tab2, tab3)
}
\arguments{
\item{tab1}{A data.table object with the \code{users.rds} table.}

\item{tab2}{A data.table object with the \code{Stats formation.xlsx} table.}

\item{tab3}{A data.table object to which you want to add this information.}
}
\value{
A data.table object with the new variable created.
}
\description{
This function will create the variable \code{formation_presentielle} which will take either 1 or 0 for value, indicating whether each individual has been involved or not in a 2-day formation.
}
\examples{
tab <- add_fp(users,infos_fp,tab)
}

