% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_proportion_mooc.R
\name{add_proportion_mooc}
\alias{add_proportion_mooc}
\title{Adds informations about the MOOC}
\usage{
add_proportion_mooc(table1, table2)
}
\arguments{
\item{table1}{The table containing all the logs.}

\item{table2}{The table containing the number of mooc used.}
}
\value{
A data.table object with the results for each indivdual.
}
\description{
This function will calculate the total proportion of MOOC chapters used for each individual.
}
\examples{
tab <- add_proportion_mooc(log,tab_mooc)
}

