% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_type_exam.R
\name{add_type_exam}
\alias{add_type_exam}
\title{Adds informations about the nature of the exam}
\usage{
add_type_exam(list_tabs, users, tab)
}
\arguments{
\item{list_tabs}{A list data.table objects with the results of exams.}

\item{users}{A data.table object with the \code{users.rds} table.}

\item{tab}{A data.table object to which you want to add the information}
}
\value{
A data.table object with the new variable created.
}
\description{
This function will create the variable \code{type_exam_passed} which will be a 3-level factor taking either ".
}
\examples{
tab <- add_type_exam(list(exams2016,exams2017),users,tab)
}

