% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_worktime.R
\name{add_worktime}
\alias{add_worktime}
\title{Adds the total time spent on the platform}
\usage{
add_worktime(log)
}
\arguments{
\item{log}{A data.table object containing the logs.}
}
\value{
A data.table object with the results wanted.
}
\description{
This function will calculate the total amount of time each individual has spent working on the platform.
}
\examples{
tab <- add_worktime(log)
}

