% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster_plots.R
\name{affiche_duration}
\alias{affiche_duration}
\title{Plots average duration of weekly work}
\usage{
affiche_duration(tab)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}
}
\value{
A bar chart.
}
\description{
This function will plot the average duration of work per week, in each cluster.
}
\examples{
affiche_duration(tab)
}

