% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_exam_attempt.R
\name{affiche_examb_att}
\alias{affiche_examb_att}
\title{Plots blank exam informations}
\usage{
affiche_examb_att(tab, phases = 5)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}

\item{phases}{A numeric indicating the number of phases. 5 is the default value.}
}
\value{
A bar chart.
}
\description{
This function will plot the proportion of individuals having attempted at least one blank exam, for each cluster.
}
\examples{
affiche_examb_att(tab)
}

