% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster_plots_eb.R
\name{affiche_examb_success}
\alias{affiche_examb_success}
\title{Plots blank exam informations}
\usage{
affiche_examb_success(tab)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}
}
\value{
A bar chart.
}
\description{
This function will plot the proportion of individuals having succeeded at the blank exam among those who hve attempted at least one, for each cluster.
}
\examples{
affiche_examb_success(tab)
}

