% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster_plots.R
\name{affiche_forma}
\alias{affiche_forma}
\title{Plots 'formation présentielle' infos}
\usage{
affiche_forma(tab, graph)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}

\item{graph.number}{A numeric indicating which graph you want to plot. 1 will plot the general participation proportion in each group. 2 will plot the results ~ participation.}
}
\value{
A bar chart.
}
\description{
This function will plot the proportion of indivudals who have been involved in the 2-day formation, in each cluster.
}
\examples{
affiche_forma(tab)
}

