% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster_plots.R
\name{affiche_mooc}
\alias{affiche_mooc}
\title{Plots MOOC informations}
\usage{
affiche_mooc(tab)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}
}
\value{
A bar chart.
}
\description{
This function will plot the proportion of MOOC chapters used and prepared in each cluster.
}
\examples{
affiche_mooc(tab)
}

