% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_mooc_cumulated.R
\name{affiche_mooc_cumulated}
\alias{affiche_mooc_cumulated}
\title{Plots MOOC informations}
\usage{
affiche_mooc_cumulated(tab, phases = 5)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}

\item{phases}{A numeric indicating the number of phases. 5 is the default value.}
}
\value{
A bar chart.
}
\description{
This function will plot the cumulated number of MOOC chapters used and prepared during each phase, for each cluster.
}
\examples{
affiche_mooc_cumulated(tab)
}

