% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_phase_duration.R
\name{affiche_phase_duration}
\alias{affiche_phase_duration}
\title{Plots duration informations}
\usage{
affiche_phase_duration(tab, phases = 5)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}

\item{phases}{A numeric indicating the number of phases. 5 is the default value.}
}
\value{
A bar chart.
}
\description{
This function will plot the average and median duration of a phase, for each cluster.
}
\examples{
affiche_phase_duration(tab)
}

