% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_progression.R
\name{affiche_progression}
\alias{affiche_progression}
\title{Plots cumulated informations}
\usage{
affiche_progression(tab, phases = 5)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}

\item{phases}{A numeric indicating the number of phases. 5 is the default value.}
}
\value{
A bar chart.
}
\description{
This function will plot the cumulated improvement and working time during each phase, for each cluster.
}
\examples{
affiche_progression(tab)
}

