% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_questions_cumulated.R
\name{affiche_questions_cumul}
\alias{affiche_questions_cumul}
\title{Plots cumulated informations}
\usage{
affiche_questions_cumul(tab, phases = 5)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}

\item{phases}{A numeric indicating the number of phases. 5 is the default value.}
}
\value{
A bar chart.
}
\description{
This function will plot the cumulated average number of questions (A and C) attempted during each phase, for each cluster.
}
\examples{
affiche_questions_cumul(tab)
}

