% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster_plots.R
\name{affiche_repartition}
\alias{affiche_repartition}
\title{Plots repartition in clusters}
\usage{
affiche_repartition(tab)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}
}
\value{
A pie chart.
}
\description{
This function will represent the percentage of individuals from the sample in each cluster, as a pie chart.
}
\examples{
affiche_repartition(tab)
}

