% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster_plots.R
\name{affiche_resultats}
\alias{affiche_resultats}
\title{Plots success rate for each cluster}
\usage{
affiche_resultats(tab)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}
}
\value{
A bar chart.
}
\description{
This function will plot the success rate at the exam for each cluster, as a bar chart.
}
\examples{
affiche_resultats(tab)
}

