% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/affiche_scores_init.R
\name{affiche_score_init}
\alias{affiche_score_init}
\title{Plots initial scores for each cluster}
\usage{
affiche_score_init(tab)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}
}
\value{
A bar chart.
}
\description{
This function will plot the initial scores after 100 questions for each cluster, as a bar chart.
}
\examples{
affiche_score_init(tab)
}

