% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ajout_origine.R
\name{affiche_suffixe}
\alias{affiche_suffixe}
\title{Extracts the origin of an individual}
\usage{
affiche_suffixe(tab1)
}
\arguments{
\item{tab1}{A data.table or data.frame object from which the email address comes from.}
}
\value{
The suffixe of the email address.
}
\description{
This function will extract the suffixe of an e-mail address.
}
\examples{
## The email is in the users table. You want to add it to the logs table
suffixe <- affiche_suffixe(users)
}

