% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster_plots_training.R
\name{affiche_tr_attempt}
\alias{affiche_tr_attempt}
\title{Plots training informations}
\usage{
affiche_tr_attempt(tab, phases = 5)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}

\item{phases}{A numeric indicating the number of phases. 5 is the default value.}
}
\value{
A bar chart.
}
\description{
This function will plot the number of training attempts in each phase, for each cluster.
}
\examples{
affiche_tr_attempt(tab)
}

