% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster_plots_quest2.R
\name{affiche_tr_quest_tot}
\alias{affiche_tr_quest_tot}
\title{Plots training informations}
\usage{
affiche_tr_quest_tot(tab, phases = 5)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}

\item{phases}{A numeric indicating the number of phases. 5 is the default value.}
}
\value{
A bar chart.
}
\description{
This function will plot the total number of questions attempted in each phase, for each cluster.
}
\examples{
affiche_tr_quest_tot(tab)
}

