% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clustering.R
\name{change_labels}
\alias{change_labels}
\title{Changes cluster labels}
\usage{
change_labels(tab)
}
\arguments{
\item{tab}{A data.table object containing the cluster labels.}
}
\value{
A data.table with the new cluster labels.
}
\description{
This function will change the cluster labels.
}
\examples{
## You perform a clustering and get these groups :

test <- clustering(tab,res_pca,nb.max.clusters = 8,size.min = 30)
summary(test$cluster)
##   1   2   3   4   6
## 184  82  34 126  25

test <- change_labels(test)
summary(test$cluster)
##   1   2   3   4   5
## 184  82  34 126  25
}

