% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/choix_seuil.R
\name{choix_seuil}
\alias{choix_seuil}
\title{Displays error rates for predictions}
\usage{
choix_seuil(obs_test, pred_test, seuils)
}
\arguments{
\item{obs_test}{A numeric or factor vector containing the observations from the test sample you try to predict.}

\item{pred_test}{A numeric vector inheriting from a \code{predict} object.}

\item{seuils}{A numeric vector with the threshold values you want to try.}
}
\value{
Plots the graph with the false positive and false negative rates for each threshold, and displays the table with the global error rate as well as false negative and false positive rates, for each threshold. Also indicates the optimal threshold
}
\description{
This function will help you when you have to choose a decision threshold to turn probabilities into binary values. Depending on your scoring problem, the false negatives and false positives haven't got the same level of importance so you can't keep a 0.5 decision threshold.
}
\note{
A dataframe with the results as well as a graph of the errors will be returned in addition.
}
\examples{
## Let's take an example with the SAheart dataset
## You can run the following code :

data("SAheart")
set.seed(234)
indapp <- sample(1:nrow(SAheart),nrow(SAheart)*(2/3),replace=FALSE)
dapp <- SAheart[indapp,]
dtest <- SAheart[-indapp,]
model_complet <- glm(chd~., data = dapp, family = binomial)
predictions <- predict(model_complet, newdata = dtest, type = "response")
seuils <- seq(0.4,0.9,by=0.05)
choix_seuil(dtest$chd,predictions,seuils)

## "Seuil minimisant le taux d'erreur global : 0.55"
}

