% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean_table.R
\name{clean_table}
\alias{clean_table}
\title{Cleans the table}
\usage{
clean_table(tab)
}
\arguments{
\item{tab}{A data.table object containing the table you want to clean.}
}
\value{
A data.table object with the results wanted.
}
\description{
This function will clean the table by removing outliers.
}
\examples{
tab_clean <- clean_table(tab)
}

