% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster_description.R
\name{cluster_description}
\alias{cluster_description}
\title{Displays cluster description}
\usage{
cluster_description(tab, cluster.label = 1)
}
\arguments{
\item{tab}{A data.table object containing the original data with cluster labes associated to each individual.}

\item{cluster.label}{A numeric indicating which cluster's characteristics you want to display. Default value is 1.}
}
\value{
The data.frame with the results will be displayed.
}
\description{
This function will display the results of the \code{v.test} on each cluster, showing variables that distinguish these individuals from others the most.
}
\examples{
## You want to see more details about the cluster with label 5
## of the dataset 'tab' ?

cluster_description(tab,cluster.label = 5)
}

