% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Performs a clustering on PCA results}
\usage{
clustering(tab, tab2, results, nb.max.clusters = 10, size.min = 30)
}
\arguments{
\item{tab}{A data.table object containing the data on which the PCA has been performed.}

\item{tab2}{A data.table object containing variables you want to add after performing the clustering for the description in each group.}

\item{results}{A list with the results of a PCA.}

\item{nb.max.clusters}{A numeric indicating the maximum number of clusters you want to have. Default value is 10.}

\item{size.min}{A numeric indicating how many individuals a cluster must contain at least. Default value is 30.}
}
\value{
A dataframe with original data and the clusters associated to each individual.
}
\description{
This function will perform a clustering from PCA results.
}
\note{
Clusters with a size inferior to what is specified inside the argument \code{size.min} will be removed.
Results will be added only after the clustering is done, as it is a non-supervised one.
}
\examples{
## You want to get a maximum of 8 clusters with at least 20 individuals in each of them :

donnees <- clustering(res_pca,nb.max.clusters = 8, size.min = 20)
}

