% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_newdata2.R
\name{create_dtest2}
\alias{create_dtest2}
\title{Creates new data to make predictions}
\usage{
create_dtest2(tab, newdata)
}
\arguments{
\item{tab}{A data.table object containing the train dataset.}

\item{newdata}{A data.table object containing the test dataset.}
}
\value{
A data.table object with the results wanted.
}
\description{
This function will create a dataframe in order to perform predictions.
}
\examples{
tab <- create_dtest2(tab,newdata)
}

