% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_matrix_pca.R
\name{create_matrix_pca}
\alias{create_matrix_pca}
\title{Creates a matrix for a PCA}
\usage{
create_matrix_pca(tab, population = "banque_detail")
}
\arguments{
\item{tab}{A data.table object containing the data you want to analyse.}

\item{population}{A character which indicates which population you want to analyse. Default value is "banque_detail", other possible value is "autre".}
}
\value{
A data frame with only numeric variables and userid as row names.
}
\description{
This function will transform your data into a matrix that can be used to realise a PCA.
}
\examples{
tab <- create_matrix_pca(tab)
}

