% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extraction_logs.R
\name{create_qr}
\alias{create_qr}
\title{Merges several tables into one}
\usage{
create_qr(tab1, tab2, tab3)
}
\arguments{
\item{tab1}{A data.table or data.frame object containing the \code{qs.rds} file.}

\item{tab2}{A data.table or data.frame object containing the \code{question.rds} file}

\item{tab3}{A data.table or data.frame object containing the \code{attempts.rds} file}
}
\value{
The final version of the table relating attempts and scores.
}
\description{
This function will create a single table of results from 3 different tables.
}
\examples{
qr <- create_qr(qs,question,attempts)
}

