% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/creation_resultats.R
\name{create_results}
\alias{create_results}
\title{Merges results}
\usage{
create_results(succes, echec, resultats_vi, users)
}
\arguments{
\item{succes}{A data.table object with the successful attempts at the AMF exam.}

\item{resultats_vi}{A data.table object with the results of the "Validation Interne" exam.}

\item{users}{A data.table object which contains informations about the userid and name of an individual.}

\item{echecs}{A data.table object with the failed attempts at the AMF exam..}
}
\value{
A unique dataset including the scores, final result, inscription and exam's date. This is returned as a data.table object.
}
\description{
This function merges results coming from different excel files into one.
}
\note{
It is important to leave spaces in the condition as the string is splitted on this pattern. Look at the next section for examples of valid conditions.

Lines with missing values will be omitted automatically.
}
\examples{
## A quick example of how to call the function :
succes <- data.table(read_excel(paste0(path,"certificats.xlsx"), sheet = "Feuil1"))
echec <- data.table(read_excel(paste0(path,"envoiEchecs.xlsm"), sheet = "Feuil1", col_names = TRUE, skip = 1))
res_vi <- data.table(read_excel(paste0(path,"Extraction Résultats Validation Interne.xlsx")))
resultats <- create_results(succes,echecs,resultats_vi)
tail(resultats,3)

## This is was you'll get :
##   userid  \%A \%C resultat         inscription     examen
##1:  49615  47 43        0 2017-04-14 10:17:56 2017-04-19
##2:  50405  88 88        1 2017-05-18 13:20:29 2011-10-01
##3:  51183 100 91        1 2017-06-15 15:55:02 2014-08-20
}

