% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/description_duration.R
\name{description_duration}
\alias{description_duration}
\title{Plots average duration of weekly work}
\usage{
description_duration(tab, phases)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}

\item{phases}{A numeric vector indicating the phases you want to plot.}
}
\value{
A bar chart.
}
\description{
This function will plot the average duration of work per week, in each group.
}
\examples{
## Display the summary of the first 3 phases
description_duration(tab,phases = 1:3)
}

