% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/description_nb_mooc_used.R
\name{description_nb_mooc_used}
\alias{description_nb_mooc_used}
\title{Plots MOOC informations}
\usage{
description_nb_mooc_used(tab, phases)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}

\item{phases}{A numeric vector indicating the phases you want to plot.}
}
\value{
A bar chart.
}
\description{
This function will plot the average number of MOOC chapters used during each phase, for each group.
}
\examples{
## Display the summary of the first 3 phases
description_nb_mooc_used(tab,phases = 1:3)
}

