% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/description_questions.R
\name{description_questions}
\alias{description_questions}
\title{Plots training informations}
\usage{
description_questions(tab, phases)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}

\item{phases}{A numeric vector indicating the phases you want to plot.}
}
\value{
A bar chart.
}
\description{
This function will plot the total number of questions attempted in each phase, for each group.
}
\examples{
## Display the summary of the first 3 phases
description_questions(tab,phases = 1:3)
}

