% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/display_pca_results.R
\name{display_pca_results}
\alias{display_pca_results}
\title{Displays PCA results}
\usage{
display_pca_results(res_pca, axe = 1)
}
\arguments{
\item{res_pca}{A list containing the PCA results.}

\item{axe}{A numeric indicating the first axis you want to use to plot the factorial plan. Default value is 1.}
}
\value{
A data.frame object with the results of the axis.
}
\description{
This function will display the table with contribution (CTR) and \code{COS²} of each individual on the axis of your choice.
}
\examples{
## You want to see the results of the first axis :

display_pca_results(res_pca, axe = 1)
}

