% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/decoupage_tables.R
\name{divide_tables}
\alias{divide_tables}
\title{Divides a table into several parts}
\usage{
divide_tables(tab1, tab2, divisions)
}
\arguments{
\item{tab1}{A data.table object containing the table you want to divide.}

\item{tab2}{A data.table  object containing the inscription date.}

\item{divisions}{A number which indicates how many parts you want to divide your table into.}
}
\value{
A list of subsets from the main table.
}
\description{
This function will divide a table into as many subsets as you want. The divisions will be made in a chronological order.
}
\examples{
## You want to divide the logs  into 5 parts

liste_logs <- divide_tables(log,res_examen,5)
}

