% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/do_pca.R
\name{do_pca}
\alias{do_pca}
\title{Performs a Principal Component Analysis}
\usage{
do_pca(tab, inertie_percentage = 95)
}
\arguments{
\item{tab}{A data.table object containing the data you want to analyse.}

\item{inertie_percentage}{A numeric between 0 and 100 indicating the percentage of variance you want to keep. Default value is 95.}
}
\value{
A list with the PCA results.
}
\description{
This function will perform a PCA on your data.
}
\examples{
## You want to keep only 85\% of variance :

res_pca <- do_pca(tab,inertie_percentage = 85)
}

