% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gap_session.R
\name{gap_session}
\alias{gap_session}
\title{Adds gaps between sessions}
\usage{
gap_session(list_tab)
}
\arguments{
\item{list_tab}{A list of the different logs tables.}
}
\value{
A data.table object with the gaps between sessions for each indivdual, across the number of phases you have chosen.
}
\description{
This function will calculate gaps between sessions (in hours) for each individual.
}
\examples{
tab <- gap_session(liste_log,divisions)
}

