% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_distance.R
\name{get_distance}
\alias{get_distance}
\title{Calculates euclidean distance}
\usage{
get_distance(tab, ind, k)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}

\item{ind}{A character indicating the userid number of the individual you want to compare to others.}

\item{k}{A numeric indicating how much nearest neighbours you want to keep.}
}
\value{
A dataset with the k-nearest neighbours.
}
\description{
This function will return the euclidean distance between an individual and others .
}
\examples{
get_distance(tab)
}

