% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_durations.R
\name{get_durations}
\alias{get_durations}
\title{Adds working time during each phase}
\usage{
get_durations(list_tab, table)
}
\arguments{
\item{list_tab}{A list of the different logs tables.}

\item{table}{A data.table object which contains the following information : the duration between the inscription and the exam.}
}
\value{
A data.table object with the results for each indivdual, across the number of phases you have chosen.
}
\description{
This function will calculate weekly and total work duration (in minutes), during each phase for each individual.
}
\examples{
tab <- get_durations(liste_log,res_examens)
}

