% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_exam_infos.R
\name{get_exam_infos}
\alias{get_exam_infos}
\title{Adds blank exams informations}
\usage{
get_exam_infos(list_tab)
}
\arguments{
\item{list_tab}{A list of tables.}
}
\value{
A data.table object with the informations wanted.
}
\description{
This function will the number of attempts, of questions attempted, and scores for each indivual accross all phases.
}
\examples{
tab <- get_exam_infos(liste_qr)
}

