% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_exam_types.R
\name{get_exam_types}
\alias{get_exam_types}
\title{Plots cluster infos}
\usage{
get_exam_types(tab)
}
\arguments{
\item{tab}{A data.table object containing the clustering dataset.}
}
\value{
A data.table object with the results for each cluster..
}
\description{
This function will calculate the proportion of people who attend each kind of exam for every cluster.
}
\examples{
res <- get_exam_types(tab)
}

