% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_mooc_infos.R
\name{get_mooc_infos}
\alias{get_mooc_infos}
\title{Adds informations about the MOOC}
\usage{
get_mooc_infos(list_tab)
}
\arguments{
\item{list_tab}{A list of the different logs tables.}
}
\value{
A data.table object with the results for each indivdual, across the number of phases you have chosen.
}
\description{
This function will calculate the number of MOOC chapters used and prepared, during each phase for each individual.
}
\examples{
tab <- gap_session(liste_log)
}

