% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_nb_questions.R
\name{get_nb_questions}
\alias{get_nb_questions}
\title{Adds the number of questions made throughout the period}
\usage{
get_nb_questions(table, phases)
}
\arguments{
\item{table}{A data.table object containing the scores.}

\item{phases}{A number indicating how many phases you have chosen to split the analysis into.}
}
\value{
A data.table object with the results for each indivdual, across the number of phases you have chosen.
}
\description{
This function will calculate the number of questions made by each individual in each phase.
}
\examples{
tab <- progression_score(table,5)
}

