% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/left_join.R
\name{left_join}
\alias{left_join}
\title{Merges tables}
\usage{
left_join(list_tab)
}
\arguments{
\item{list_tab}{A list of the tables you want to merge.}
}
\value{
A unique data.table object which will be the result of the \code{LEFT JOIN}.
}
\description{
This function will perform a \code{LEFT JOIN} on the userid key.
}
\examples{
tab <- left_join(liste_tabs)
}

