% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/replace_na.R
\name{na_replace}
\alias{na_replace}
\title{Replaces NA in a data.table}
\usage{
na_replace(DT, value)
}
\arguments{
\item{DT}{A data.table object.}

\item{value}{A length-one numeric, character or factor vector indicating which value you want to use to replace \code{NAs} in your data.table.}
}
\value{
The data.table without NAs.
}
\description{
This function will replace NAs in the data.table with the value of your choice.
}
\examples{
## You want to replace all you NAs by 0

DT <- na_replace(DT,0)
}

