% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot.clusters}
\alias{plot.clusters}
\title{Plots clusters's size}
\usage{
plot.clusters(tab)
}
\arguments{
\item{tab}{A data.table or data.frame object containing the data with clusters.}
}
\value{
A dataframe with original data and the clusters associated to each individual.
}
\description{
This function will perform a barplot to represent each cluster's size.
}

