% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_cumulated_time.R
\name{plot_cumulated_time}
\alias{plot_cumulated_time}
\title{Plots cumulated working time}
\usage{
plot_cumulated_time(tab, phases = 5)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}

\item{phases}{A numeric indicating the number of phases. 5 is the default value.}
}
\value{
A bar chart.
}
\description{
This function will plot the cumulated working time (in hours) during each phase, for each cluster.
}
\examples{
plot_cumulated_time(tab)
}

