% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/progression_score.R
\name{progression_score}
\alias{progression_score}
\title{Adds scores progressions through the working period}
\usage{
progression_score(table, phases)
}
\arguments{
\item{table}{A data.table object containing the scores.}

\item{phases}{A number indicating how many phases you have chosen to split the analysis into.}
}
\value{
A data.table object with the results for each indivdual, across the number of phases you have chosen.
}
\description{
This function will calculate score progressions for each level (A and C). The result will be a percentage.
}
\examples{
tab <- progression_score(table,5)
}

