% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/do_pca.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Removes outliers from a PCA}
\usage{
remove_outliers(tab, results, min.inertie = 5, ctr.min = 2)
}
\arguments{
\item{tab}{A data.table object containing the data you want to analyse.}

\item{results}{A list with the results of a PCA.}

\item{min.inertie}{A numeric between 0 and 100 indicating the minimum iniertia an axis must have to be taken into consideration. Default value is 5}

\item{ctr.min}{A numeric indicating the CTR threshold above which you consider that an individual is an outlier. Default value is 2.}
}
\value{
A data.table object without outliers.
}
\description{
This function will remove outliers from your analysis.
}
\examples{
## You want to remove all the indivuals that have a CTR > 1.90 on all the
## axis which represent at least 4\% of variance :

tab_clean <- remove_outliers(tab,res_pca,min.inertie = 4, ctr.min = 1.90)
}

