% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/select_modelisation.R
\name{select_modelisation}
\alias{select_modelisation}
\title{Selects variables}
\usage{
select_modelisation(tab)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}
}
\value{
A data.table object with the data to analyse.
}
\description{
This function will select variables from the general dataset, to keep only the first 3 phases informations.
}
\examples{
select_modelisation(tab)
}

