% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/select_var_tree.R
\name{select_var_tree}
\alias{select_var_tree}
\title{Selects variables}
\usage{
select_var_tree(tab)
}
\arguments{
\item{tab}{A data.table object containing the clustering data.}
}
\value{
A data.table object with the data to analyse.
}
\description{
This function will select variables from the general dataset to include them in the decision tree.
}
\examples{
select_var_tree(tab)
}

