% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/choix_seuil.R
\name{vec_assign}
\alias{vec_assign}
\title{Assigns binary values to a vector}
\usage{
vec_assign(vecteur, seuil = 0.5)
}
\arguments{
\item{vecteur}{A numeric vector to transform.}

\item{seuil}{A numeric indicating the decision threshold. 0.5 is the default value.}
}
\value{
A vector with 0 and 1 values.
}
\description{
This function will assign 0 if value < seuil, 1 otherwise.
}
\examples{
vector <- c(0.2,0.6,0.75,0.05)
vector <- vec_assign(vector,0.6)
print(vector)
## [1] 0 1 1 0
}

