% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/display_pca_results.R
\name{display_pca_results}
\alias{display_pca_results}
\title{Displays PCA results}
\usage{
display_pca_results(res_pca, axe1 = 1, axe2 = 2)
}
\arguments{
\item{res_pca}{A list containing the PCA results.}

\item{axe1}{A numeric indicating the first axis you want to use to plot the factorial plan. Default value is 1.}

\item{axe2}{A numeric indicating the second axis you want to use to plot the factorial plan. Default value is 2.}
}
\value{
Plots the factorial plan you chose as well as PCA results for these axis.
}
\description{
This function will plot the factorial plan of your choice as well as the table with contribution (CTR) and \code{COS²} of each individual on the correspondant axis.
}
\examples{
## You want to see the results on the first factorial plan

display_pca_results(res_pca, axe1 = 1, axe2 = 2)
}

