% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_initial_level.R
\name{get_initial_level}
\alias{get_initial_level}
\title{Adds informations about the initial level}
\usage{
get_initial_level(list_tab, id = NULL)
}
\arguments{
\item{list_tab}{A list of tables.}

\item{id}{A character indicating the userid you want to create. Default value is \code{NULL}.}
}
\value{
A data.table object with the results for each indivdual.
}
\description{
This function will calculate the initial level in terms of scores after 100 questions for each individual.
}
\examples{
tab <- get_initial_level(liste_qr)
}

